Refactored mockdata and trending page
- Separated mock data from component code
- Added trending page to include navbar
- Edited that when click to Trending page, navigate to 'trending'
