+++ frontend/newsbridge/src/pages/Trending.tsx
@@ -1,95 +1,18 @@
 import ArticlePreview from "../components/Article/ArticlePreview";
-import { UnbiasedArticlePreviewViewModel } from "../entities/viewmodels/UnbiasedArticlePreviewVM";
 import { useState } from "react";
+import mockArticles from "../mocks/mockArticles";
 
-const mockArticles: UnbiasedArticlePreviewViewModel[] = [
-  {
-    id: 1,
-    title: "This is a Title For a News Story",
-    summary: "This is a summary of a news article that is being displayed on the home page. This is a summary of a news article that is being displayed on the home page. This is a summary of a news article that is being displayed on the home page. This is a summary of a news article that is being displayed on the home page.",
-    content: "Full content of the article",
-    imageUrl: "/",
-    topic: "US News",
-    likeCount: 10,
-    genre: "News",
-    createdAt: "2025-10-10",
-    daysAgo: "3 days ago",
-    isBookmarked: true,
-    biasRating: 50,
-    commentCount: 13,
-  },
-  {
-    id: 2,
-    title: "This is a Title For a News Story",
-    summary: "This is a summary of a news article that is being displayed on the home page. This is a summary of a news article that is being displayed on the home page. This is a summary of a news article that is being displayed on the home page. This is a summary of a news article that is being displayed on the home page.",
-    content: "Full content of the article",
-    imageUrl: "/",
-    topic: "US News",
-    likeCount: 10,
-    genre: "News",
-    createdAt: "2025-10-10",
-    daysAgo: "3 days ago",
-    isBookmarked: true,
-    biasRating: 50,
-    commentCount: 19,
-  },
-  {
-    id: 3,
-    title: "This is a Title For a News Story",
-    summary: "This is a summary of a news article that is being displayed on the home page. This is a summary of a news article that is being displayed on the home page. This is a summary of a news article that is being displayed on the home page. This is a summary of a news article that is being displayed on the home page.",
-    content: "Full content of the article",
-    imageUrl: "/",
-    topic: "US News",
-    likeCount: 10,
-    genre: "News",
-    createdAt: "2025-10-10",
-    daysAgo: "3 days ago",
-    isBookmarked: true,
-    biasRating: 50,
-    commentCount: 22,
-  },
-  {
-    id: 4,
-    title: "This is a Title For a News Story",
-    summary: "This is a summary of a news article that is being displayed on the home page. This is a summary of a news article that is being displayed on the home page. This is a summary of a news article that is being displayed on the home page. This is a summary of a news article that is being displayed on the home page.",
-    content: "Full content of the article",
-    imageUrl: "/",
-    topic: "US News",
-    likeCount: 10,
-    genre: "News",
-    createdAt: "2025-10-10",
-    daysAgo: "3 days ago",
-    isBookmarked: true,
-    biasRating: 50,
-    commentCount: 10,
-  },
-  {
-    id: 5,
-    title: "This is a Title For a News Story",
-    summary: "This is a summary of a news article that is being displayed on the home page. This is a summary of a news article that is being displayed on the home page. This is a summary of a news article that is being displayed on the home page. This is a summary of a news article that is being displayed on the home page.",
-    content: "Full content of the article",
-    imageUrl: "/",
-    topic: "US News",
-    likeCount: 10,
-    genre: "News",
-    createdAt: "2025-10-10",
-    daysAgo: "3 days ago",
-    isBookmarked: true,
-    biasRating: 50,
-    commentCount: 10,
-  },
-]
 const itemPerPage = 3
 const Trending = () => {
   const [curPage, setCurPage] = useState(1);
   const totalPages = Math.ceil(mockArticles.length / itemPerPage);
-
 
   const handlePageChange = (newPage: number) => {
     if(newPage >=1 && newPage <= totalPages){
       setCurPage(newPage)
     }
   }
+  
   const paginatedArticles = mockArticles.slice(
     (curPage - 1) * itemPerPage,
     curPage * itemPerPage
